11-lorawan: Remove task 06 GNRC LoRaWAN (ABP)
After some discussion it would seem like the best option to handle the generally untested task is to just remove it.

This has not been testing is quite some time, likey due to the somewhat challenging setup requirering chipstack.
The alternative would be to acutally implement the missing operations, however, the ABP generally is not used
in production, so we would be testing something only used for testing... which seems a bit redundant.